+++ src/main/java/com/example/rolebase/repository/UserRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     @Query("SELECT u FROM User u " + "LEFT JOIN FETCH u.roles ur " +
             "LEFT JOIN FETCH ur.role " + "WHERE LOWER(u.username) = LOWER(:username)")
@@ -21,7 +21,6 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     boolean existsByEmail(String email);
 
     @Modifying
-    @Query("UPDATE User u SET u.enabled = :enabled " +
-            "WHERE LOWER(u.username) = LOWER(:username)")
+    @Query("UPDATE User u SET u.enabled = :enabled " + "WHERE LOWER(u.username) = LOWER(:username)")
     int updateUserEnabledStatus(@Param("username") String username, @Param("enabled") boolean enabled);
 }
